HSEARCH-2418 Move analyzer definition translation code to a specific package
This will make room for classes with a different purpose in the
"settings.impl" package.
