+++ elasticsearch/src/main/java/org/hibernate/search/elasticsearch/settings/impl/ElasticsearchIndexSettingsBuilder.java
@@ -20,6 +20,7 @@ import org.hibernate.search.elasticsearch.settings.impl.model.CharFilterDefiniti
 import org.hibernate.search.elasticsearch.settings.impl.model.IndexSettings;
 import org.hibernate.search.elasticsearch.settings.impl.model.TokenFilterDefinition;
 import org.hibernate.search.elasticsearch.settings.impl.model.TokenizerDefinition;
+import org.hibernate.search.elasticsearch.settings.impl.translation.ElasticsearchAnalyzerDefinitionTranslator;
 import org.hibernate.search.engine.spi.EntityIndexBinding;
 import org.hibernate.search.util.logging.impl.LoggerFactory;
