+++ elasticsearch/src/main/java/org/hibernate/search/elasticsearch/settings/impl/translation/AnalysisDefinitionFactory.java
@@ -4,10 +4,14 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.elasticsearch.settings.impl;
+package org.hibernate.search.elasticsearch.settings.impl.translation;
 
-import com.google.gson.JsonElement;
+import org.hibernate.search.annotations.Parameter;
 
-interface ParameterValueTransformer {
-	JsonElement transform(String parameterValue);
+interface AnalysisDefinitionFactory<D> {
+
+	String getType();
+
+	D create(Parameter[] parameters);
+
 }
