+++ elasticsearch/src/main/java/org/hibernate/search/elasticsearch/settings/impl/translation/ParameterValueTransformer.java
@@ -4,14 +4,10 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.elasticsearch.settings.impl;
+package org.hibernate.search.elasticsearch.settings.impl.translation;
 
-import org.hibernate.search.annotations.Parameter;
+import com.google.gson.JsonElement;
 
-interface AnalysisDefinitionFactory<D> {
-
-	String getType();
-
-	D create(Parameter[] parameters);
-
+interface ParameterValueTransformer {
+	JsonElement transform(String parameterValue);
 }
