+++ elasticsearch/src/main/java/org/hibernate/search/elasticsearch/settings/impl/translation/StemmerOverrideRuleFileParameterValueTransformer.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.elasticsearch.settings.impl;
+package org.hibernate.search.elasticsearch.settings.impl.translation;
 
 import java.io.IOException;
 import java.io.InputStream;
