--- elasticsearch/src/main/resources/META-INF/services/org.hibernate.search.elasticsearch.settings.impl.ElasticsearchAnalyzerDefinitionTranslator
@@ -1,1 +1,0 @@
-org.hibernate.search.elasticsearch.settings.impl.DefaultElasticsearchAnalyzerDefinitionTranslator
