+++ elasticsearch/src/main/resources/META-INF/services/org.hibernate.search.elasticsearch.settings.impl.translation.ElasticsearchAnalyzerDefinitionTranslator
@@ -1,0 +1,1 @@
+org.hibernate.search.elasticsearch.settings.impl.translation.DefaultElasticsearchAnalyzerDefinitionTranslator
