+++ elasticsearch/src/test/java/org/hibernate/search/elasticsearch/test/DefaultElasticsearchAnalyzerDefinitionTranslatorTest.java
@@ -30,10 +30,10 @@ import org.hibernate.search.annotations.TokenFilterDef;
 import org.hibernate.search.annotations.TokenizerDef;
 import org.hibernate.search.elasticsearch.analyzer.ElasticsearchTokenFilterFactory;
 import org.hibernate.search.elasticsearch.impl.JsonBuilder;
-import org.hibernate.search.elasticsearch.settings.impl.DefaultElasticsearchAnalyzerDefinitionTranslator;
 import org.hibernate.search.elasticsearch.settings.impl.model.CharFilterDefinition;
 import org.hibernate.search.elasticsearch.settings.impl.model.TokenFilterDefinition;
 import org.hibernate.search.elasticsearch.settings.impl.model.TokenizerDefinition;
+import org.hibernate.search.elasticsearch.settings.impl.translation.DefaultElasticsearchAnalyzerDefinitionTranslator;
 import org.hibernate.search.exception.SearchException;
 import org.hibernate.search.testsupport.setup.BuildContextForTest;
 import org.hibernate.search.testsupport.setup.SearchConfigurationForTest;
